requirements.txt
torch
torchvision
streamlit
diffusers
transformers
Pillow
qrcode
requests